out-of-line shmem/TransportFactory.h
